Updated Functionality to Spline
Added multiple functionalities to the spline tool.
Now when adding a cubic curve to the spline, we can enforce a constraint upon the new curve. We can decide whether we want the new spline to have neither a first or second derivative, in which case no constraint is enforced. We can decide if we want the new spline to have a first derivative, in which case an alignment constraint is enforced. Or we can decide if we want to have both a first and second derivative, in which case a mirroring constraint is enforced. 
For the purpose of our line integral module, we will want to have the second derivative, but the additional functionality was free to add on.

Additionally, I have added a looping functionality. We can have the spline loop back onto itself. A closed loop path is necessary for amply explaining line integrals.