Walker
Added a component that walks along a given Bezier curve and faces in the direction of the curve. We need to replace the walker component with an arrow. This arrow will point in the direction of the derivative along the curve, and in the context of a line integral module, would be the vector that would be dotted with the field at the given position.